plugin: prefer plugins loaded later than the ones loaded early
This patch changes the preference of the plugin framework to prefer
the one that is loaded later, instead of the one loaded earlier.

The reason for the change, is that blocks would start using the plugin
framework (e.g. all user defined blocks would become dynamically
allocated plugins) and the behaviour there is to prefer the one that
comes last.

Since, with normal plugins we don't expect collisions, for them the change
in behavior is inmaterial.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>